requirements.txt
asgiref==3.3.4
Django==3.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
mysqlclient==1.4.6
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1